fixing incoming and outgoing update quantity in item interaction and add wishlist

+++ frontend/src/components/incomingitemlist.jsx
@@ -51,18 +51,24 @@ const Incomingitemlist = () => {
 
     const deleteIncomingItem = async (itemId, iuid) => {
 
-        await axios.delete(`http://localhost:5000/incomingItems/${itemId}`);
-        await axios.patch("http://localhost:5000/updateQuantityOnHand", {
-            iuid: iuid
-        });
-        const newTotalRows = rows - 1;
-        setItems(incomingitems.filter((item) => item.id !== itemId));
-        setRows(newTotalRows);
-        const newTotalPages = Math.ceil(newTotalRows / limit);
-        if (page >= newTotalPages) {
-            setPage(newTotalPages - 1);
+        const confirmed = window.confirm('Are you sure you want to delete this item?');
+        if (confirmed) {
+            await axios.delete(`http://localhost:5000/incomingItems/${itemId}`);
+            await axios.patch("http://localhost:5000/updateQuantityReceived", {
+                iuid: iuid
+            });
+            await axios.patch("http://localhost:5000/updateQuantityOnHand", {
+                iuid: iuid
+            });
+            const newTotalRows = rows - 1;
+            setItems(incomingitems.filter((item) => item.id !== itemId));
+            setRows(newTotalRows);
+            const newTotalPages = Math.ceil(newTotalRows / limit);
+            if (page >= newTotalPages) {
+                setPage(newTotalPages - 1);
+            }
+            setPages(newTotalPages);
         }
-        setPages(newTotalPages);
     };
